move functions to branch from reqs
